// src/app/journey/models/follow.model.ts
export class FollowUp {

    code: string;
    notes: string;
    nextActionAt: string;
    reminder: string;
    total: number;
    ownerUniqueId: string;
    ownerName: string;
    ownerEmail: string;
    duration: number;
    durationUnit: number;
    durationDescription: string;
    status: string;
    enabled: string;

}